# docs/swagger.yaml
basePath: /
definitions:
...
  commonModel.Meta:
    properties:
      page:
        example: 1
        type: integer
      perPage:
        example: 10
        type: integer
      totalItems:
        example: 100
        type: integer
      totalPages:
        example: 10
        type: integer
    type: object
...
  disbursementModel.Disbursement:
    properties:
      amount:
        example: 1000000
        type: number
      approvedAt:
        example: "2021-01-01T00:00:00Z"
        type: string
      approvedBy:
        example: Lorem Ipsum
        type: string
      bankReferenceNo:
        example: WErft7089990
        type: string
      beneficiaryAccountName:
        example: Yories Yolanda
        type: string
      beneficiaryAccountNo:
        example: "8000800808"
        type: string
      beneficiaryBankCode:
        example: "008"
        type: string
      beneficiaryBankName:
        example: Bank Permata
        type: string
      bulkId:
        example: ff35c906-0144-4a52-8621-4991dd00f945
        type: string
      createdAt:
        example: "2021-01-01T00:00:00Z"
        type: string
      createdBy:
        example: John Doe
        type: string
      createdFrom:
        example: MERCHANT_PORTAL | OPEN_API
        type: string
      currency:
        example: IDR
        type: string
      deletedAt:
        example: "2021-01-01T00:00:00Z"
        type: string
      fee:
        example: 4000
        type: number
      merchantId:
        example: 50f95963-373a-4a6e-a7c9-79b6cf891f9c
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/disbursementModel.Metadata'
        description: Internal Data
      processorReferenceId:
        example: dce9c63e-8a1c-4d8f-b18d-33de6f772104
        type: string
      purposeId:
        example: bad7bd6f-3459-4e28-97d5-e200469d5cc6
        type: string
      reasonDescription:
        example: User input
        type: string
      reasonType:
        example: OTHER
        type: string
      referenceId:
        example: reference-id-from-client
        type: string
      remark:
        example: Disburse to Yories
        type: string
      senderName:
        example: Google Corp
        type: string
      status:
        example: APPROVED
        type: string
      totalAmount:
        example: 1004000
        type: number
      updatedAt:
        example: "2021-01-01T00:00:00Z"
        type: string
      uuid:
        example: b7b95825-6ec3-486a-a3cf-62d4ed815757
        type: string
    type: object
...
  disbursementModel.Metadata:
    properties:
      feeDetail:
        $ref: '#/definitions/feeModel.FeeMetadataObject'
    type: object
...
  feeModel.FeeMetadataObject:
    properties:
      amount:
        type: number
      amountType:
        type: string
      deductionType:
        type: string
      linked_transaction_id:
        description: For fee that have a parent transaction recorded in the account_transactions
          table
        type: string
      method:
        type: string
      percentage:
        type: number
      taxAmount:
        type: number
      taxPercentage:
        type: number
      taxType:
        type: string
      type:
        type: string
    type: object
...
  merchant.AddressDetail:
    properties:
      city:
        type: string
      district:
        type: string
      province:
        type: string
    type: object
  merchant.BoardOfDirectorResp:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      identityFile:
        type: string
      identityNumber:
        type: string
      name:
        type: string
      positionLong:
        type: string
      updatedAt:
        type: string
    type: object
...
  merchant.ListBoardOfDirectorResp:
    properties:
      commissionerList:
        items:
          $ref: '#/definitions/merchant.BoardOfDirectorResp'
        type: array
      directorList:
        items:
          $ref: '#/definitions/merchant.BoardOfDirectorResp'
        type: array
    type: object
...
  merchant.MerchantRequest:
    properties:
      address:
        maxLength: 254
        type: string
      businessCountry:
        type: string
      businessStructure:
        type: string
      businessType:
        type: string
      description:
        type: string
      districtId:
        minimum: 1
        type: integer
      logo:
        type: string
      merchantEmail:
        type: string
      merchantPhone:
        type: string
      name:
        type: string
      parentID:
        type: string
      picEmail:
        type: string
      picJobTitle:
        type: string
      picName:
        type: string
      picPhone:
        type: string
      postcode:
        type: string
      requesterID:
        type: string
      requesterType:
        type: string
      shortName:
        maxLength: 25
        type: string
      userId:
        type: string
    required:
    - address
    - businessCountry
    - businessStructure
    - businessType
    - districtId
    - logo
    - merchantEmail
    - merchantPhone
    - name
    - picEmail
    - picName
    - picPhone
    - postcode
    - shortName
    type: object
  merchant.MerchantResponse:
    properties:
      active:
        type: boolean
      address:
        type: string
      addressDetail:
        $ref: '#/definitions/merchant.AddressDetail'
      businessCountry:
        type: string
      businessStructure:
        type: string
      businessType:
        type: string
      createdAt:
        type: string
      description:
        type: string
      districtId:
        type: integer
      externalId:
        type: string
      logo:
        type: string
      merchantEmail:
        type: string
      merchantPhone:
        type: string
      name:
        type: string
      parentId:
        type: string
      picEmail:
        type: string
      picJobTitle:
        type: string
      picName:
        type: string
      picPhone:
        type: string
      postcode:
        type: string
      shortName:
        type: string
      updatedAt:
        type: string
      uuid:
        type: string
    type: object
...
  merchant.UpsertBoardOfDirectorReq:
    type: object
  merchant.UpsertBoardOfDirectorResp:
    properties:
      id:
        type: string
    type: object
...
  response.ApiError:
    properties:
      details:
        items:
          $ref: '#/definitions/response.ApiErrorDetail'
        type: array
      traceId:
        type: string
      type:
        type: string
    type: object
  response.ApiErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.ApiResponse:
    properties:
      code:
        example: "00"
        type: string
      data: {}
      error:
        $ref: '#/definitions/response.ApiError'
      message:
        type: string
      pagination: {}
    type: object
...
  response.Response:
    properties:
      code:
        example: "00"
        type: string
      data: {}
      errors: {}
      meta: {}
    type: object
...
host: localhost
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    To support payment gateway needs, we prepare technical documents on the backend-portal.
    Where the backend-portal is intended to serve services to OpenAPI and MerchantPortal,
    apart from that this service is integrated with Orchestrator, Core Processor
    (SNAP Core Processor and Credit Card Processor)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Portal API
  version: "1.0"
paths:
...
  /api/v1/disbursements:
    get:
      consumes:
      - application/json
      description: Disbursement list endpoint
      operationId: api-disbursement-list
      parameters:
      - description: pagination current page
        in: query
        name: page
        type: string
      - description: filter bulkId
        in: query
        name: bulkId
        type: string
      - description: filter status
        in: query
        name: status
        type: string
      - description: filter startCreatedAt
        in: query
        name: startCreatedAt
        type: string
      - description: filter endCreatedAt
        in: query
        name: endCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/disbursementModel.Disbursement'
                  type: array
                meta:
                  $ref: '#/definitions/commonModel.Meta'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - Bearer: []
      summary: Disbursement list endpoint
      tags:
      - API - Disbursement
...
  /api/v1/disbursements/{id}/receipt:
    get:
      consumes:
      - application/json
      description: Disbursement Receipt by ID endpoint
      operationId: api-disbursement-receipt-by-id
      parameters:
      - description: Disbursement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/disbursementModel.Disbursement'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - Bearer: []
      summary: Disbursement Receipt by ID endpoint
      tags:
      - API - Disbursement
...
  /api/v1/merchants/create:
    post:
      consumes:
      - application/json
      description: Create merchant endpoint
      operationId: merchant-create
      parameters:
      - description: JSON Body for Create Merchant
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/merchant.MerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/merchant.MerchantResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - Bearer: []
      summary: Create merchant endpoint
      tags:
      - API - Merchant
...
  /crm/v1/merchants/{id}/bods:
    get:
      consumes:
      - application/json
      description: Endpoint for get list data Board Of Director (BOD)
      operationId: merchant-get-list-bod
      parameters:
      - description: Merchant Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-CRM-Key:
              description: '{"key": "value"}'
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/merchant.ListBoardOfDirectorResp'
              type: object
        "500":
          description: Internal Server Error
          headers:
            X-CRM-Key:
              description: '{"key": "value"}'
              type: string
          schema:
            $ref: '#/definitions/response.Response'
      summary: Endpoint for get list data Board Of Director (BOD)
      tags:
      - API - Merchant Board Of Directors
    post:
      consumes:
      - multipart/form-data
      description: Endpoint for create data Board Of Director (BOD)
      operationId: merchant-create-bod
      parameters:
      - description: Merchant Id
        in: path
        name: id
        required: true
        type: string
      - description: JSON Body for create board of director data
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/merchant.UpsertBoardOfDirectorReq'
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          headers:
            X-CRM-Key:
              description: '{"key": "value"}'
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/merchant.UpsertBoardOfDirectorResp'
              type: object
        "500":
          description: Internal Server Error
          headers:
            X-CRM-Key:
              description: '{"key": "value"}'
              type: string
          schema:
            $ref: '#/definitions/response.Response'
      summary: Endpoint for create data Board Of Director (BOD)
      tags:
      - API - Merchant Board Of Directors
  /crm/v1/merchants/{id}/bods/{bod_id}:
    put:
      consumes:
      - multipart/form-data
      description: Endpoint for update data Board Of Director (BOD)
      operationId: merchant-update-bod
      parameters:
      - description: Merchant Id
        in: path
        name: id
        required: true
        type: string
      - description: Unique Id for Board Of Director (BOD)
        in: path
        name: bod_id
        required: true
        type: string
      - description: JSON Body for update board of director data
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/merchant.UpsertBoardOfDirectorReq'
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          headers:
            X-CRM-Key:
              description: '{"key": "value"}'
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/merchant.UpsertBoardOfDirectorResp'
              type: object
        "500":
          description: Internal Server Error
          headers:
            X-CRM-Key:
              description: '{"key": "value"}'
              type: string
          schema:
            $ref: '#/definitions/response.Response'
      summary: Endpoint for update data Board Of Director (BOD)
      tags:
      - API - Merchant Board Of Directors
...
schemes:
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"